Remove run.h

+++ 2_lexical_fuzzing/1_fuzzing_braking_thing_with_random_inputs/fuzzing_architecture/include/fuzz_input.h
@@ -3,6 +3,6 @@
 
 #include "config.h"
 
-int_fast16_t fuzz_input (test_config_t * config, char * fuzz_string) ;
+int fuzz_input (test_config_t * config, char * fuzz_string) ;
 
 #endif
